Wrap timeline image into a div element and style it for tiny screens
The added div element with `timeline-wrapper` class allows easy styling
of the content. Add a new module, too, which is added to the output, whenever
a timeline is rendered and sets the max-width to 100% with an automatic over-
flow, so timelines, which are bigger as the content div (e.g. on tiny screens)
can scroll the timeline without breaking the content style.

Bug: T118601
Change-Id: I09c5b168e02a9a3aff30d59671986cfc086cf8c9