Migrate to BoxedCommand, copy structure from Score
The main goal of this patch is to migrate to BoxedCommand so this
extension can be run with Shellbox. Most of the structure has been
copied from the Score extension, specifically 2dba77355d by Tim
Starling, who has given permission to relicense and reuse it in this
extension. Unifying with Score will hopefully make this extension more
maintainable going forward.

A shell wrapper script, scripts/renderTimeline.sh, wraps around
scripts/EasyTimeline.pl to generate the SVG or PNG. If a SVG is
generated, rsvg will be used to rasterize to PNG. An alternative SVG
coverter can be used, but it needs to be command-line compatible with
rsvg-convert for now.

The Docker image configuration for Wikimedia usage is available
at I3c1fea59c60.

* Refactor error handling to use a TimelineException, similar to
  ScoreException. To faciliate this, split renderTimeline into an outer
  onTagHook function, which checks if rendering is needed, calls
  renderTimeline(), and catches TimelineException and formats it.
* Use ScopedCallback to remove the created temporary directory instead
  of relying on an implicit destructor that isn't as obvious to
  developers.
* Fonts can now be in multiple directories and specified by users with
  e.g. <timeline font="noto">. $wgTimelineFonts should contain keys
  that users can use and values to the full path of the font. Note that
  the default should not be changed as most timelines are pixel aligned
  to the current font.
* Build cache key in the same way as Score, introduce a CACHE_VERSION
  constant and drop $wgTimelineEpochTimestamp and use of
  $wgRenderHashAppend. Adding ploticus version to the cache key has been
  skipped for now since it's dead upstream.
* Use Html methods to replace nearly all manual HTML construction.
* Remove obsolete comments about generating gifs, was dropped in
  0ca732c0cb2682 (2012).
* Drop .timeline-error CSS replicating <tt>, we now wrap the appropriate
  stuff in <pre>.

Bug: T289226
Change-Id: I5ab357f17617a4f9c1a33ff08ec812ed41714745